more or less having success with the callback. Still haven't tried running, no idea if it works. not sure what I'm going to do about having to pass in a shared pref value into the presenter.
