+++ app/src/main/java/net/rmoreno/weatherapp/MainActivity.java
@@ -19,16 +19,9 @@ import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
-
-import com.squareup.okhttp.Call;
-import com.squareup.okhttp.Callback;
-import com.squareup.okhttp.OkHttpClient;
-import com.squareup.okhttp.Request;
-import com.squareup.okhttp.Response;
 
 import net.rmoreno.weatherapp.adapters.DailyAdapter;
 import net.rmoreno.weatherapp.models.CurrentWeather;
@@ -38,7 +31,6 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Map;
 
@@ -116,6 +108,10 @@ public class MainActivity extends Activity implements DailyWeatherView{
     }
 
     @Override
+    public void displayDailyWeather() {
+
+    }
+    @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
 
@@ -127,48 +123,9 @@ public class MainActivity extends Activity implements DailyWeatherView{
 
     }
 
-    public CurrentWeather getCurrentWeatherData(String jsonData) throws JSONException{
 
-        CurrentWeather currentWeather = new CurrentWeather();
-        JSONObject jsonObject = new JSONObject(jsonData);
 
-        JSONObject currently = jsonObject.getJSONObject("currently");
 
-        Log.d(ACTIVITY + "currently", currently.toString());
-
-        currentWeather.setTemp(currently.getDouble("temperature"));
-        currentWeather.setSummary(currently.getString("summary"));
-        currentWeather.setTime(currently.getLong("time"));
-        currentWeather.setTimeZone(jsonObject.getString("timezone"));
-        currentWeather.setIcon(currently.getString("icon"));
-        currentWeather.setPrecip(currently.getInt("precipProbability"));
-        currentWeather.setFeels(currently.getDouble(("apparentTemperature")));
-        currentWeather.setWind(currently.getDouble(("windSpeed")));
-
-        return currentWeather;
-    }
-
-    public ArrayList<DailyWeather> getDailyWeatherData(String jsonData) throws JSONException {
-        ArrayList<DailyWeather> dailyWeatherList = new ArrayList<>();
-
-        JSONObject jsonObject = new JSONObject(jsonData);
-        JSONObject hourly = jsonObject.getJSONObject("daily");
-        JSONArray data = hourly.getJSONArray("data");
-
-        Log.d(ACTIVITY, String.valueOf(data.getJSONObject(1).getDouble("precipProbability")));
-        for (int i = 0; i < data.length(); i++) {
-            DailyWeather dailyWeather = new DailyWeather();
-
-            dailyWeather.setTime(data.getJSONObject(i).getLong("time"));
-            dailyWeather.setMinTemp(data.getJSONObject(i).getDouble("temperatureMin"));
-            dailyWeather.setMaxTemp(data.getJSONObject(i).getDouble("temperatureMax"));
-            dailyWeather.setTimeZone(jsonObject.getString("timezone"));
-            dailyWeather.setPrecip(data.getJSONObject(i).getDouble("precipProbability"));
-            dailyWeatherList.add(dailyWeather);
-        }
-
-        return dailyWeatherList;
-    }
 
     public void displayDailyWeather(CurrentWeather current, ArrayList<DailyWeather> daily, int sweaterTemp) {
         temperature.setText(current.getTemp() + "°");
